ScheduleRestController en bijhorende testen

+++ src/test/java/be/kdg/runtracker/frontend/controllers/rest/UserRestControllerTest.java
@@ -7,7 +7,6 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.MockitoAnnotations;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
@@ -41,34 +40,40 @@ public class UserRestControllerTest {
     private MockMvc mockMvc;
     private Gson gson;
 
+    private User alexander;
+    private User wout;
+    private User jelle;
+    private User stijn;
+    private User jens;
+
     @Before
     public void setup() {
 
-        User alexander = new User();
+        alexander = new User();
         alexander.setAuthId(123);
         alexander.setFirstname("Alexander");
         alexander.setLastname("van Ravestyn");
         alexander.setUsername("alexvr");
 
-        User wout = new User();
+        wout = new User();
         wout.setAuthId(234);
         wout.setFirstname("Wout");
         wout.setLastname("Lambreghts");
         wout.setUsername("woutl");
 
-        User jelle = new User();
+        jelle = new User();
         jelle.setAuthId(345);
         jelle.setFirstname("Jelle");
         jelle.setLastname("Mannaerts");
         jelle.setUsername("jellem");
 
-        User stijn = new User();
+        stijn = new User();
         stijn.setAuthId(456);
         stijn.setFirstname("Stijn");
         stijn.setLastname("Ergeerts");
         stijn.setUsername("stijne");
 
-        User jens = new User();
+        jens = new User();
         jens.setAuthId(567);
         jens.setFirstname("Jens");
         jens.setLastname("Schadron");
@@ -80,7 +85,6 @@ public class UserRestControllerTest {
         userRepository.save(stijn);
         userRepository.save(jens);
 
-        MockitoAnnotations.initMocks(this);
         this.mockMvc = webAppContextSetup(webApplicationContext).build();
         this.gson = new Gson();
 
@@ -91,6 +95,24 @@ public class UserRestControllerTest {
         mockMvc.perform(get("/api/users").contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andDo(print());
+    }
+
+    @Test
+    public void testGetNoUsers() throws Exception {
+        userRepository.delete(userRepository.findUserByAuthId(123).getUser_id());
+        userRepository.delete(userRepository.findUserByAuthId(234).getUser_id());
+        userRepository.delete(userRepository.findUserByAuthId(345).getUser_id());
+        userRepository.delete(userRepository.findUserByAuthId(456).getUser_id());
+        userRepository.delete(userRepository.findUserByAuthId(567).getUser_id());
+
+        mockMvc.perform(get("/api/users").contentType(MediaType.APPLICATION_JSON))
+                .andExpect(status().isNoContent());
+
+        userRepository.save(alexander);
+        userRepository.save(wout);
+        userRepository.save(jelle);
+        userRepository.save(stijn);
+        userRepository.save(jens);
     }
 
     @Test
@@ -178,10 +200,12 @@ public class UserRestControllerTest {
         if (userRepository.findUserByAuthId(123) != null) {
             userRepository.delete(userRepository.findUserByAuthId(123).getUser_id());
         }
-        userRepository.delete(userRepository.findUserByAuthId(234).getUser_id());
-        userRepository.delete(userRepository.findUserByAuthId(345).getUser_id());
-        userRepository.delete(userRepository.findUserByAuthId(456).getUser_id());
-        userRepository.delete(userRepository.findUserByAuthId(567).getUser_id());
+        if (!userRepository.findAll().isEmpty()) {
+            userRepository.delete(userRepository.findUserByAuthId(234).getUser_id());
+            userRepository.delete(userRepository.findUserByAuthId(345).getUser_id());
+            userRepository.delete(userRepository.findUserByAuthId(456).getUser_id());
+            userRepository.delete(userRepository.findUserByAuthId(567).getUser_id());
+        }
 
     }
